Update community-health-assessment-philadelphia-department-of-public-health.md
The URL @andrewviren provided is a good addition. However, it is a completely different data product that has different contents that do not match the description or the metadata entry, and I think pointing people to a the Health of the City Report from this catalog entry would be a mistake. I've updated the entry with the Health Explorer link. I think it would be best to:

1) determine if Public Health plans to fix the Community Health Explorer, and, if not, delete the catalog entry
2) create a new entry from the Health of the City Report

I will create issues for both of these.