Add non-blocking action to validate data.json against DCAT schema
Uses an external NASA validator under the hood: https://github.com/nasa/DCATUS-Validator

The action logs out any validation results to aid in getting our data.json file up to spec. To see the validation results, hit "Checks" on any PR, go to the dcat job, and expand the "Run dcat-validator" step. At this point, this will include some validations that don't apply to us, like the bureauCode, so the action is always non-blocking for a merge. The intention is just for it to be a useful tool for quickly checking the results of any changes we make to `data.json`.

This PR also modifies data.json for a simplified alternative solution to https://github.com/opendataphilly/opendataphilly-jkan/pull/412 meant to address the specific issues with URLs noted by the validator, i.e. spaces and pipes.